requirements.txt
python-telegram-bot

# for .env file support
python-dotenv~=0.20.0